feat: separate rules
